fix(fw): Fix Blocks generator, add assertion on ModelGenerator
