+++ src/ethereum_test_tools/tests/test_generators.py
@@ -8,31 +8,7 @@ from typing import List
 import pytest
 
 from ..common import Address, Transaction, Transactions, Withdrawal, Withdrawals
-from ..common.base_types import ModelGenerator
-from ..spec.blockchain.types import Block
-
-
-# Blocks is currently broken, single test involving Blocks is expected to fail
-class Blocks(ModelGenerator, model=Block):
-    """
-    Block generator.
-
-    This class is used to generate blocks for a blockchain test case.
-
-    Takes the same arguments as `Block`, but if an iterable is provided for a
-    field, it will generate a block for each element in the iterable.
-
-    If none of the values are iterables, an exception will be raised because none of the provided
-    values will be bounded, unless the `limit` argument is provided, which will limit the number
-    of blocks generated.
-
-    Fields that are already supposed to be iterables, such as `txs` and
-    `withdrawals`, can be parametrized as `txs_iter` and
-    `withdrawals_iter` respectively and an iterable of iterables can be provided to be
-    used in each generated block.
-    """
-
-    pass
+from ..spec.blockchain.types import Block, Blocks
 
 
 @pytest.mark.parametrize(
@@ -560,7 +536,12 @@ class Blocks(ModelGenerator, model=Block):
         ),
         pytest.param(
             Blocks(
-                txs_iter=Transactions(
+                # One quirk of the ModelGenerator is that, when used as a field value to other
+                # ModelGenerators, it has to be added to the normal field instead of
+                # the `_iter` field.
+                # Reason being is that the iterator is generated again on every iteration,
+                # so the generator is embedded into a `cycle` function.
+                txs=Transactions(
                     to=(Address(i * 0x100) for i in count(1)),
                     limit=[3, 2],
                 ),
@@ -575,9 +556,11 @@ class Blocks(ModelGenerator, model=Block):
                                 to=Address(0x100),
                             ),
                             Transaction(
+                                nonce=1,
                                 to=Address(0x200),
                             ),
                             Transaction(
+                                nonce=2,
                                 to=Address(0x300),
                             ),
                         ],
@@ -585,16 +568,17 @@ class Blocks(ModelGenerator, model=Block):
                     Block(
                         txs=[
                             Transaction(
+                                nonce=3,
                                 to=Address(0x400),
                             ),
                             Transaction(
+                                nonce=4,
                                 to=Address(0x500),
                             ),
                         ],
                     ),
                 ]
             ],
-            marks=pytest.mark.xfail(reason="Blocks generator is broken"),
             id="blocks-1",
         ),
     ],
@@ -604,6 +588,9 @@ def test_generators(
     chunks: int,
     expected_lists: List[List[Transaction | Withdrawal | Block]],
 ):
+    """
+    Test generating models using the provided model generators.
+    """
     chunked_elements: List[List[Transaction | Withdrawal | Block]] = []
     for _ in range(chunks):
         chunked_elements.append(list(generator))
